move responsibility of pattern path creation out of the PatternPiece and into the Document. Moved functions for creating allowances into AllowanceFinder.
